refactor: consolidate audit logging into domain-specific modules (closes #114)
