+++ lib/authify_web/controllers/saml_providers_controller.ex
@@ -1,9 +1,9 @@
 defmodule AuthifyWeb.SAMLProvidersController do
   use AuthifyWeb, :controller
 
-  alias Authify.AuditLog
   alias Authify.SAML
   alias Authify.SAML.ServiceProvider
+  alias AuthifyWeb.Audit.SAML, as: SAMLAudit
 
   # Safely convert string to atom, only for known valid values
   defp safe_to_atom(string)
@@ -136,24 +136,8 @@ defmodule AuthifyWeb.SAMLProvidersController do
 
   # Helper for audit logging SAML service providers
   defp log_saml_provider_event(conn, event_type, service_provider, metadata) do
-    organization = conn.assigns.current_organization
-    current_user = conn.assigns.current_user
-
-    AuditLog.log_event_async(event_type, %{
-      organization_id: organization.id,
-      actor_type: "user",
-      actor_id: current_user.id,
-      actor_name: "#{current_user.first_name} #{current_user.last_name}",
-      resource_type: "saml_service_provider",
-      resource_id: service_provider.id,
-      outcome: "success",
-      ip_address: to_string(:inet_parse.ntoa(conn.remote_ip)),
-      user_agent: Plug.Conn.get_req_header(conn, "user-agent") |> List.first(),
-      metadata:
-        Map.merge(metadata, %{
-          service_provider_name: service_provider.name,
-          service_provider_id: service_provider.id
-        })
-    })
+    SAMLAudit.log_saml_provider_event(conn, event_type, service_provider,
+      extra_metadata: metadata
+    )
   end
 end
